semantic-similarity-pipeline/requirements.txt
PyYAML
fpdf2
sentence-transformers
scikit-learn
arxiv
PyMuPDF
numpy